+++ src/test/kotlin/com/pswidersk/gradle/python/TestUtils.kt
@@ -1,0 +1,5 @@
+package com.pswidersk.gradle.python
+
+import java.io.File
+
+fun loadResource(location: String): File = File(object {}.javaClass.classLoader.getResource(location)!!.file)