+++ CHANGELOG.md
@@ -6,9 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-### Changed
-
-### Removed
+- Properly mapping submodules in the SDK Import config.
 
 ## 2.6.0 - 2024-04-20